test(index): split writer verifier coverage
